agentic_chatbot/requirements.txt
flask==3.1.0
phidata
groq
googlesearch-python
yfinance
python-dotenv
requests
pycountry
gunicorn